+++ src/test/java/blackjack/domain/ParticipantsTest.java
@@ -1,16 +1,8 @@
 package blackjack.domain;
 
-import static blackjack.domain.Number.ACE;
-import static blackjack.domain.Number.FOUR;
-import static blackjack.domain.Number.THREE;
-import static blackjack.domain.Number.TWO;
-import static blackjack.domain.Symbol.DIAMOND;
-import static blackjack.domain.Symbol.HEART;
-import static blackjack.domain.Symbol.SPADE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import java.util.Map;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -44,26 +36,5 @@ class ParticipantsTest {
         for (List<Card> cards : playersCards) {
             assertThat(cards).hasSize(2);
         }
-    }
-
-    @DisplayName("플레이어의 보유 카드를 모두 확인한다")
-    @Test
-    void should_OpenCards_Of_AllPlayers() {
-        List<String> playerNames = List.of("odo", "doy");
-        Participants participants = Participants.of(playerNames);
-
-        DeckGenerator mockGenerator = new MockDeckGenerator((List.of(
-                new Card(SPADE, ACE), new Card(SPADE, TWO),
-                new Card(DIAMOND, THREE), new Card(DIAMOND, FOUR),
-                new Card(HEART, THREE), new Card(HEART, FOUR)
-        )));
-        participants.handInitialCards(mockGenerator.generate());
-
-        Map<String, List<Card>> cardsByParticipants = participants.openPlayerCards();
-        assertThat(cardsByParticipants)
-                .containsAllEntriesOf(Map.of(
-                        "odo", List.of(new Card(DIAMOND, THREE), new Card(DIAMOND, FOUR)),
-                        "doy", List.of(new Card(HEART, THREE), new Card(HEART, FOUR))
-                ));
     }
 }
